Separate estinants from voictents and update file names to match
