+++ packages/voictents-and-estinants-engine/src/custom/programmable-units/engine-program/tree/engineProgramTreeNode.ts
@@ -1,14 +1,7 @@
-import { Tuple } from '../../../../utilities/semantic-types/tuple';
-import { buildEstinant } from '../../../adapter/estinant-builder/estinantBuilder';
 import { Grition } from '../../../adapter/grition';
 import { OdeshinFromGrition } from '../../../adapter/odeshin';
 import { Voictent } from '../../../adapter/voictent';
-import { EngineProgramVoictent, ENGINE_PROGRAM_GEPP } from '../engineProgram';
-import {
-  EstinantTreeNode,
-  EstinantTreeNodeVoictent,
-  ESTINANT_TREE_NODE_GEPP,
-} from './estinantTreeNode';
+import { EstinantTreeNode } from './estinantTreeNode';
 
 export type EngineProgramEstinantTreeNode = Pick<
   EstinantTreeNode,
@@ -33,32 +26,3 @@ export type EngineProgramTreeNodeVoictent = Voictent<
   EngineProgramTreeNodeGepp,
   EngineProgramTreeNodeOdeshin
 >;
-
-export const constructEngineProgramTree = buildEstinant()
-  .fromGrition<EngineProgramVoictent>({
-    gepp: ENGINE_PROGRAM_GEPP,
-  })
-  .andFromHubblepupTuple<EstinantTreeNodeVoictent, Tuple<string>>({
-    gepp: ESTINANT_TREE_NODE_GEPP,
-    framate: (leftInput) => leftInput.grition.estinantIdentifierList,
-    croard: (rightInput) => rightInput.zorn,
-  })
-  .toGrition<EngineProgramTreeNodeVoictent>({
-    gepp: ENGINE_PROGRAM_TREE_NODE_GEPP,
-    getZorn: (leftInput) => leftInput.zorn,
-  })
-  .onPinbe((engineProgram, rightInputList) => {
-    const estinantTreeNodeList = rightInputList.map(({ grition }) => grition);
-
-    return {
-      programName: engineProgram.programName,
-      estinantList: estinantTreeNodeList.map<EngineProgramEstinantTreeNode>(
-        ({ estinantName, inputList, outputList }) => ({
-          estinantName,
-          inputList,
-          outputList,
-        }),
-      ),
-    };
-  })
-  .assemble();
