+++ packages/voictents-and-estinants-engine/src/custom/programmable-units/engine-program/tree/estinantTreeNode.ts
@@ -1,21 +1,8 @@
-import { buildEstinant } from '../../../adapter/estinant-builder/estinantBuilder';
 import { Grition } from '../../../adapter/grition';
 import { OdeshinFromGrition } from '../../../adapter/odeshin';
 import { Voictent } from '../../../adapter/voictent';
-import {
-  EstinantInputOutputParentVoictent,
-  ESTINANT_INPUT_OUTPUT_PARENT_GEPP,
-} from '../estinant-call-expression-parameter/estinantInputOutputParent';
-import {
-  EstinantInput,
-  EstinantInputListVoictent,
-  ESTINANT_INPUT_LIST_GEPP,
-} from '../estinant-input-output/estinantInputList';
-import {
-  EstinantOutput,
-  EstinantOutputListVoictent,
-  ESTINANT_OUTPUT_LIST_GEPP,
-} from '../estinant-input-output/estinantOutputList';
+import { EstinantInput } from '../estinant-input-output/estinantInputList';
+import { EstinantOutput } from '../estinant-input-output/estinantOutputList';
 
 export type EstinantTreeInputNode = Pick<
   EstinantInput,
@@ -44,38 +31,3 @@ export type EstinantTreeNodeVoictent = Voictent<
   EstinantTreeNodeGepp,
   EstinantTreeNodeOdeshin
 >;
-
-export const constructEstinantTreeNode = buildEstinant()
-  .fromGrition<EstinantInputOutputParentVoictent>({
-    gepp: ESTINANT_INPUT_OUTPUT_PARENT_GEPP,
-  })
-  .andFromHubblepupTuple<EstinantInputListVoictent, [string]>({
-    gepp: ESTINANT_INPUT_LIST_GEPP,
-    framate: (leftInput) => [leftInput.grition.inputListIdentifier],
-    croard: (rightInput) => rightInput.zorn,
-  })
-  .andFromHubblepupTuple<EstinantOutputListVoictent, [string]>({
-    gepp: ESTINANT_OUTPUT_LIST_GEPP,
-    framate: (leftInput) => [leftInput.grition.outputListIdentifier],
-    croard: (rightInput) => rightInput.zorn,
-  })
-  .toGrition<EstinantTreeNodeVoictent>({
-    gepp: ESTINANT_TREE_NODE_GEPP,
-    getZorn: (leftInput) => leftInput.zorn,
-  })
-  .onPinbe(
-    (estinantParent, [{ grition: inputList }], [{ grition: outputList }]) => {
-      return {
-        programName: estinantParent.programName,
-        estinantName: estinantParent.estinantName,
-        inputList: inputList.map(({ voictentName, index }) => ({
-          voictentName,
-          index,
-        })),
-        outputList: outputList.map(({ voictentName }) => ({
-          voictentName,
-        })),
-      };
-    },
-  )
-  .assemble();
