+++ packages/voictents-and-estinants-engine/src/custom/programmable-units/type-script-file/parsedTypeScriptFile.ts
@@ -1,15 +1,7 @@
 import { TSESTree } from '@typescript-eslint/typescript-estree';
-import * as parser from '@typescript-eslint/typescript-estree';
-import fs from 'fs';
 import { Grition } from '../../adapter/grition';
 import { OdeshinFromGrition } from '../../adapter/odeshin';
-import {
-  TypeScriptFileConfigurationVoictent,
-  TYPE_SCRIPT_FILE_CONFIGURATION_GEPP,
-} from './typeScriptFileConfiguration';
-import { ErrorVoictent, ERROR_GEPP } from '../error/error';
 import { Voictent } from '../../adapter/voictent';
-import { buildEstinant } from '../../adapter/estinant-builder/estinantBuilder';
 
 export type ParsedTypeScriptFile = {
   filePath: string;
@@ -29,51 +21,3 @@ export type ParsedTypeScriptFileVoictent = Voictent<
   ParsedTypeScriptFileGepp,
   ParsedTypeScriptFileOdeshin
 >;
-
-export const parseTypeScriptFile = buildEstinant()
-  .fromHubblepup<TypeScriptFileConfigurationVoictent>({
-    gepp: TYPE_SCRIPT_FILE_CONFIGURATION_GEPP,
-  })
-  .toHubblepupTuple<ParsedTypeScriptFileVoictent>({
-    gepp: PARSED_TYPE_SCRIPT_FILE_GEPP,
-  })
-  .toHubblepupTuple<ErrorVoictent>({
-    gepp: ERROR_GEPP,
-  })
-  .onPinbe((input) => {
-    const inputGrition = input.grition;
-
-    const fileContents = fs.readFileSync(inputGrition.sourceFilePath, 'utf8');
-
-    try {
-      const program: TSESTree.Program = parser.parse(fileContents, {
-        project: './tsconfig.json',
-        tsconfigRootDir: inputGrition.rootDirectory,
-        comment: true,
-      });
-
-      return {
-        [PARSED_TYPE_SCRIPT_FILE_GEPP]: [
-          {
-            zorn: input.zorn,
-            grition: {
-              filePath: inputGrition.sourceFilePath,
-              program,
-            },
-          },
-        ],
-        [ERROR_GEPP]: [],
-      };
-    } catch (error) {
-      return {
-        [PARSED_TYPE_SCRIPT_FILE_GEPP]: [],
-        [ERROR_GEPP]: [
-          {
-            zorn: input.zorn,
-            grition: error,
-          },
-        ],
-      };
-    }
-  })
-  .assemble();
