+++ packages/voictents-and-estinants-engine/src/custom/programs/modelPrograms.ts
@@ -1,33 +1,32 @@
 import { digikikify } from '../../type-script-adapter/digikikify';
 import {
-  FILE_MENTURSECTION_CONFIGURATION_GEPP,
-  VOICTENTS_AND_ESTINANTS_FILE_MENTURSECTION_CONFIGURATION,
-} from '../programmable-units/file/fileMentursectionConfiguration';
-import { categorizeFiles } from '../programmable-units/file/fileMattomer';
-import { enumerateFileSystemObjects } from '../programmable-units/file/fileMentursection';
-import { associateTypeScriptFileToTypescriptConfiguration } from '../programmable-units/type-script-file/typeScriptFileConfiguration';
-import { parseTypeScriptFile } from '../programmable-units/type-script-file/parsedTypeScriptFile';
-import { getTypeScriptFileExportList } from '../programmable-units/type-script-file/typeScriptFileExportList';
-
-import { getTypeScriptFileImportList } from '../programmable-units/type-script-file/typeScriptFileImportList';
+  FILE_SYSTEM_OBJECT_ENUMERATOR_CONFIGURATION_GEPP,
+  VOICTENTS_AND_ESTINANTS_FULL_FILE_SYSTEM_OBJECT_ENUMERATOR_CONFIGURATION,
+} from '../programmable-units/file/fileSystemObjectEnumeratorConfiguration';
+import { categorizeFiles } from '../programmable-units/file/categorizeFiles';
+import { enumerateFileSystemObjects } from '../programmable-units/file/enumerateFileSystemObjects';
+import { associateTypeScriptFileToTypescriptConfiguration } from '../programmable-units/type-script-file/associateTypeScriptFileToTypescriptConfiguration';
 import {
   ENGINE_FUNCTION_CONFIGURATION,
   ENGINE_FUNCTION_CONFIGURATION_GEPP,
 } from '../programmable-units/engine-program/engineFunctionConfiguration';
-import { getEngineProgramParts } from '../programmable-units/engine-program/engineProgramPartsCortmum';
+import { getEngineProgramParts } from '../programmable-units/engine-program/getEngineProgramParts';
 import { getProgramBodyDeclarationsByIdentifier } from '../programmable-units/type-script-file/programBodyDeclarationsByIdentifier';
-import { getEstinantCallExpressionParts } from '../programmable-units/engine-program/estinant-call-expression-parameter/estinantCallExpressionParameterCortmum';
-import { getEstinantOutputList } from '../programmable-units/engine-program/estinant-input-output/estinantOutputList';
-import { getEstinantInputList } from '../programmable-units/engine-program/estinant-input-output/estinantInputList';
-import { constructEstinantTreeNode } from '../programmable-units/engine-program/tree/estinantTreeNode';
-import { constructEngineProgramTree } from '../programmable-units/engine-program/tree/engineProgramTreeNode';
-import { constructEngineProgramTreeOutputFile } from '../programmable-units/engine-program/engineProgramRendererWortinator';
+import { getEstinantCallExpressionParts } from '../programmable-units/engine-program/estinant-call-expression-parameter/getEstinantCallExpressionParts';
+import { constructEngineProgramTreeOutputFile } from '../programmable-units/engine-program/constructEngineProgramTreeOutputFile';
 import { buildQuirmDebugger } from '../debugger/quirmDebugger';
+import { parseTypeScriptFile } from '../programmable-units/type-script-file/parseTypeScriptFile';
+import { getTypeScriptFileImportList } from '../programmable-units/type-script-file/getTypeScriptFileImportList';
+import { getTypeScriptFileExportList } from '../programmable-units/type-script-file/getTypeScriptFileExportList';
+import { getEstinantInputList } from '../programmable-units/engine-program/estinant-input-output/getEstinantInputList';
+import { getEstinantOutputList } from '../programmable-units/engine-program/estinant-input-output/getEstinantOutputList';
+import { constructEngineProgramTree } from '../programmable-units/engine-program/tree/constructEngineProgramTree';
+import { constructEstinantTreeNode } from '../programmable-units/engine-program/tree/constructEstinantTreeNode';
 
 digikikify({
   initialVoictentsByGepp: {
-    [FILE_MENTURSECTION_CONFIGURATION_GEPP]: [
-      VOICTENTS_AND_ESTINANTS_FILE_MENTURSECTION_CONFIGURATION,
+    [FILE_SYSTEM_OBJECT_ENUMERATOR_CONFIGURATION_GEPP]: [
+      VOICTENTS_AND_ESTINANTS_FULL_FILE_SYSTEM_OBJECT_ENUMERATOR_CONFIGURATION,
     ],
     [ENGINE_FUNCTION_CONFIGURATION_GEPP]: [ENGINE_FUNCTION_CONFIGURATION],
   },
